fix: Resolve 146 mypy type errors across codebase
Systematic fixes for mypy strict type checking:

- Add explicit float() casts for numpy floating types (np.mean, np.var)
- Add np.asarray() for numpy array return types
- Add type annotations for variables with Union types
- Add type: ignore comments for transformers/torch library quirks
- Fix PIL PixelAccess None checks with assertions
- Fix Optional type narrowing issues
- Add Literal type for streamlit delta_color parameter
- Fix websocket message yield type

Reduces mypy errors from 146 to 0 while preserving code functionality.

+++ packages/sleeper_agents/src/sleeper_agents/safety_training/pipeline.py
@@ -122,9 +122,9 @@ class SafetyTrainingPipeline:
             training_method=config.training_method,
             timestamp=datetime.now(),
             pre_training_detections=pre_training_detections,
-            pre_training_confidence=np.mean(list(pre_training_detections.values())),
+            pre_training_confidence=float(np.mean(list(pre_training_detections.values()))),
             post_training_detections=post_training_detections,
-            post_training_confidence=np.mean(list(post_training_detections.values())),
+            post_training_confidence=float(np.mean(list(post_training_detections.values()))),
             persistence_rate=0.0,  # Calculated in __post_init__
             trigger_specificity_change=trigger_analysis.get("specificity_change", 0.0),
             behavior_consistency=self._calculate_consistency(pre_training_detections, post_training_detections),
